# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanji's All Blue Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PhoneInput {
  phone: string;
  name?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  phone: string;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResult {
  success: boolean;
  user: User;
}

export type MenuItemOrigin = typeof MenuItemOrigin[keyof typeof MenuItemOrigin];


export const MenuItemOrigin = {
  north_blue: 'north_blue',
  south_blue: 'south_blue',
  east_blue: 'east_blue',
  west_blue: 'west_blue',
} as const;

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  origin: MenuItemOrigin;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
}

export type MenuItemInputOrigin = typeof MenuItemInputOrigin[keyof typeof MenuItemInputOrigin];


export const MenuItemInputOrigin = {
  north_blue: 'north_blue',
  south_blue: 'south_blue',
  east_blue: 'east_blue',
  west_blue: 'west_blue',
} as const;

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  origin: MenuItemInputOrigin;
  category: string;
  imageUrl?: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
}


export interface Chef {
  id: number;
  name: string;
  specialty: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  yearsExperience?: number;
}

export interface Review {
  id: number;
  /** @nullable */
  menuItemId?: number | null;
  userName?: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  menuItemId?: number;
  rating: number;
  comment: string;
  userName?: string;
}

export interface OrderItem {
  menuItemId: number;
  name?: string;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  preparing: 'preparing',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  reference?: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  userId?: number | null;
  status: OrderStatus;
  total: number;
  deliveryAddress: string;
  /** @nullable */
  deliveryLat?: number | null;
  /** @nullable */
  deliveryLng?: number | null;
  /** @nullable */
  estimatedMinutes?: number | null;
  items?: OrderItem[];
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export interface OrderInput {
  userId?: number;
  deliveryAddress: string;
  deliveryLat?: number;
  deliveryLng?: number;
  items: OrderItem[];
}

export type AdminOrderUpdateStatus = typeof AdminOrderUpdateStatus[keyof typeof AdminOrderUpdateStatus];


export const AdminOrderUpdateStatus = {
  pending: 'pending',
  preparing: 'preparing',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface AdminOrderUpdate {
  status?: AdminOrderUpdateStatus;
  /** @nullable */
  adminNote?: string | null;
}

export type AdminUserInputRole = typeof AdminUserInputRole[keyof typeof AdminUserInputRole];


export const AdminUserInputRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface AdminUserInput {
  /** @nullable */
  name?: string | null;
  phone?: string;
  role?: AdminUserInputRole;
}

export type DashboardStatsOriginBreakdown = {[key: string]: number};

export interface DashboardStats {
  totalMenuItems: number;
  totalChefs: number;
  totalReviews: number;
  averageRating: number;
  totalOrders: number;
  originBreakdown?: DashboardStatsOriginBreakdown;
}

export type ListMenuItemsParams = {
origin?: ListMenuItemsOrigin;
category?: string;
search?: string;
};

export type ListMenuItemsOrigin = typeof ListMenuItemsOrigin[keyof typeof ListMenuItemsOrigin];


export const ListMenuItemsOrigin = {
  north_blue: 'north_blue',
  south_blue: 'south_blue',
  east_blue: 'east_blue',
  west_blue: 'west_blue',
} as const;

export type ListReviewsParams = {
menuItemId?: number;
};

export type AdminDeleteMenuItem200 = {
  success?: boolean;
};

export type AdminDeleteOrder200 = {
  success?: boolean;
};

export type AdminDeleteReview200 = {
  success?: boolean;
};

export type AdminDeleteUser200 = {
  success?: boolean;
};