# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sanji's All Blue Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerPhone
      tags: [auth]
      summary: Register with mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/login:
    post:
      operationId: loginPhone
      tags: [auth]
      summary: Login with mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhoneInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /chefs/{id}:
    get:
      operationId: getChef
      tags: [chefs]
      summary: Get a chef
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chef
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chef"
...
  /admin/menu/{id}:
    put:
      operationId: adminUpdateMenuItem
      tags: [admin]
      summary: Update a menu item (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteMenuItem
      tags: [admin]
      summary: Delete a menu item (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
  /admin/orders/{id}:
    patch:
      operationId: adminUpdateOrder
      tags: [admin]
      summary: Update order status and admin note (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteOrder
      tags: [admin]
      summary: Delete an order (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/reviews:
    get:
      operationId: adminListReviews
      tags: [admin]
      summary: List all reviews (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: All reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        "403":
          description: Forbidden
  /admin/reviews/{id}:
    delete:
      operationId: adminDeleteReview
      tags: [admin]
      summary: Delete a review (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all registered users (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
  /admin/users/{id}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update a user account (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user account (admin only)
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "403":
          description: Forbidden
        "404":
          description: Not found
components:
  schemas:
...
    PhoneInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
        name:
          type: string
    AuthResult:
      type: object
      required: [success, user]
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, phone, role]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, admin]
        createdAt:
          type: string
    MenuItem:
      type: object
      required: [id, name, price, origin, category]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        origin:
          type: string
          enum: [north_blue, south_blue, east_blue, west_blue]
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
    MenuItemInput:
      type: object
      required: [name, price, origin, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        origin:
          type: string
          enum: [north_blue, south_blue, east_blue, west_blue]
        category:
          type: string
        imageUrl:
          type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
    Chef:
      type: object
      required: [id, name, specialty]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        yearsExperience:
          type: integer
    Review:
      type: object
      required: [id, rating, comment, createdAt]
      properties:
        id:
          type: integer
        menuItemId:
          type: ["integer", "null"]
        userName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
...
    OrderItem:
      type: object
      required: [menuItemId, quantity, price]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
    Order:
      type: object
      required: [id, reference, status, total, deliveryAddress, createdAt]
      properties:
        id:
          type: integer
        reference:
          type: string
        customerName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, preparing, out_for_delivery, delivered, cancelled]
        total:
          type: number
        deliveryAddress:
          type: string
        deliveryLat:
          type: ["number", "null"]
        deliveryLng:
          type: ["number", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminOrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, out_for_delivery, delivered, cancelled]
        adminNote:
          type: ["string", "null"]
    AdminUserInput:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: string
        role:
          type: string
          enum: [customer, admin]
...