# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanji's All Blue Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserInputRole } from './adminUserInputRole';

export interface AdminUserInput {
  /** @nullable */
  name?: string | null;
  phone?: string;
  role?: AdminUserInputRole;
}