# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanji's All Blue Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  status: OrderStatus;
  total: number;
  deliveryAddress: string;
  /** @nullable */
  deliveryLat?: number | null;
  /** @nullable */
  deliveryLng?: number | null;
  /** @nullable */
  estimatedMinutes?: number | null;
  items?: OrderItem[];
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}