# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sanji's All Blue Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  phone: string;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  createdAt?: string;
}